requirements.txt
streamlit
pillow==9.3.0
numpy
diffusers==0.3.0
transformers
ftfy
rembg